Move linear_extrude outside teeth loop
Performing a linear_extrude on each tooth of a gear is both inefficient and prone to generating invalid geometry with internal surfaces. By moving the operation outside the loop the gear profile is assembled fully in 2D before extruding.

Before this change the geometry generated by the example has 236 CSG elements, and is very slow to manipulate in the preview window. After the change it has just 18 elements, and is correspondingly much faster.

I believe this change has no negative effects.